requirements.txt
django-allauth
django-htmx
crispy-bootstrap5
django-crispy-forms
django-filter
pillow